requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
google-api-python-client==2.125.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.30.0
aiosmtplib==3.0.1
jinja2==3.1.4
pydantic==2.7.0
httpx==0.27.0
gspread==6.1.0
